Call solve from part 1 and part 2.

+++ 2024/day20/day20.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 namespace day20 {
-    int part1(const std::vector<std::string>& rows, int min_cheat = 100);
+    size_t part1(const std::vector<std::string>& rows, int min_cheat = 100);
     size_t part2(const std::vector<std::string>& rows, int min_cheat = 100);
 
     size_t solve(const std::vector<std::string>& rows, int min_cheat, int cheat_length);
